+++ heap.c
@@ -1,5 +1,6 @@
 #include <stdint.h>
 #include <stdio.h>
+#include <inttypes.h>
 
 #include "allocator.h"
 #include "values.h"
@@ -7,14 +8,14 @@
 #include "display.h"
 
 static Data *heap = NULL;
-static uint64_t hp = 0;
+static uint32_t hp = 0;
 
 #define heap_check() if(address > hp){ \
-        printf(error("Bad heap access : %ld"), address); \
+        printf(error("Bad heap access : %" PRIu32), address); \
         return 0;}
 
-uint64_t heap_add_int(int32_t i){
-    uint64_t start = 0;
+uint32_t heap_add_int(int32_t i){
+    uint32_t start = 0;
     while(start < hp){
         if(isint(heap[start]) && tint(heap[start]) == i)
             return start;
@@ -25,8 +26,8 @@ uint64_t heap_add_int(int32_t i){
     return hp - 1;
 }
 
-uint64_t heap_add_float(double i){
-    uint64_t start = 0;
+uint32_t heap_add_float(double i){
+    uint32_t start = 0;
     while(start < hp){
         if(isfloat(heap[start]) && tfloat(heap[start]) == i)
             return start;
@@ -37,8 +38,8 @@ uint64_t heap_add_float(double i){
     return hp - 1;
 }
 
-uint64_t heap_add_str(uint64_t key){
-    uint64_t start = 0;
+uint32_t heap_add_str(uint32_t key){
+    uint32_t start = 0;
     while(start < hp){
         if(isstr(heap[start]) && tstrk(heap[start]) == key)
             return start;
@@ -49,8 +50,8 @@ uint64_t heap_add_str(uint64_t key){
     return hp - 1; 
 }
 
-uint64_t heap_add_identifer(uint64_t key){
-    uint64_t start = 0;
+uint32_t heap_add_identifer(uint32_t key){
+    uint32_t start = 0;
     while(start < hp){
         if(isidentifer(heap[start]) && tstrk(heap[start]) == key)
             return start;
@@ -61,8 +62,8 @@ uint64_t heap_add_identifer(uint64_t key){
     return hp - 1; 
 }
 
-uint64_t heap_add_logical(int32_t i){
-    uint64_t start = 0;
+uint32_t heap_add_logical(int32_t i){
+    uint32_t start = 0;
     while(start < hp){
         if(islogical(heap[start]) && tint(heap[start]) == i)
             return start;
@@ -73,30 +74,30 @@ uint64_t heap_add_logical(int32_t i){
     return hp - 1;
 }
 
-Data heap_get_data(uint64_t address){
+Data heap_get_data(uint32_t address){
     if(address > hp - 1){
-        printf(error("Invalid heap access : %lu"), address);
+        printf(error("Invalid heap access : %" PRIu32 ), address);
         return new_null();
     }
     return heap[address];
 }
 
-int32_t heap_get_int(uint64_t address){
+int32_t heap_get_int(uint32_t address){
     heap_check();
     return tint(heap[address]);
 }
 
-double heap_get_float(uint64_t address){
+double heap_get_float(uint32_t address){
     heap_check();
     return tfloat(heap[address]);
 }
 
-uint64_t heap_get_str(uint64_t address){
+uint32_t heap_get_str(uint32_t address){
     heap_check();
     return tstrk(heap[address]);
 }
 
-int32_t heap_get_logical(uint64_t address){
+int32_t heap_get_logical(uint32_t address){
     heap_check();
     return tint(heap[address]);
 }
