+++ values.c
@@ -6,21 +6,21 @@
 #include "env.h"
 #include "display.h"
 
-static uint64_t id = 0;
+static uint32_t id = 0;
 
-Data new_array(uint64_t size){
+Data new_array(uint32_t size){
     Data d;
     d.type = ARR;
     d.refCount = 0;
     d.numElements = size;
     d.arr = (Data *)mallocate(sizeof(Data)*size);
-    uint64_t i = 0;
+    uint32_t i = 0;
     while(i < size)
         d.arr[i++] = new_null();
     return d;
 }
 
-Data new_ins(void *env, uint64_t name){
+Data new_ins(void *env, uint32_t name){
     Data d;
     d.type = INSTANCE;
     d.pvalue = (Instance *)mallocate(sizeof(Instance));
@@ -57,7 +57,7 @@ void data_free(Data d){
         }
     }
     else if(isarray(d)){
-        int64_t i = 0;
+        uint32_t i = 0;
         Data *arr = (Data *)d.arr;
         while(i < d.numElements){
             data_free(arr[i]);
